Solution to Input and Print an Integer

+++ Input and Print an Integer@@6788d5012db25efec946f939/code.c
@@ -2,8 +2,8 @@
 
 
 int main (){
-    double a;
-    scanf("%lf",&a);
-    printf("You entered; %lf", a);
+    int a;
+    scanf("%d",&a);
+    printf("You entered; %d", a);
     return 0
 }
